remove binding for .rocksdb_property_int/.rocksdb_property_int_cf
they are not exported by rocksdb on windows
