+++ db_test.go
@@ -13,9 +13,6 @@ func TestOpenDb(t *testing.T) {
 	db := newTestDB(t, nil)
 	defer db.Close()
 	require.EqualValues(t, "0", db.GetProperty("rocksdb.num-immutable-mem-table"))
-	v, success := db.GetIntProperty("rocksdb.num-immutable-mem-table")
-	require.EqualValues(t, uint64(0), v)
-	require.True(t, success)
 }
 
 func TestDBCRUD(t *testing.T) {
